fix the problem with force converting the statically renders routes to dynamic
