+++ src/app/(main)/main/@modal/default.tsx
@@ -1,3 +1,5 @@
+import CanvaPlaceHolder from "@/components/CanvaPlaceHolder";
+
 export default function Default() {
-  return null;
+  return <CanvaPlaceHolder />;
 }
