+++ src/app/(main)/main/page.tsx
@@ -1,30 +1,4 @@
-"use client";
-
 import SideNav from "@/components/SideNav";
-import { readFlowsFromDatabase } from "@/lib/api-controllers";
-import useFlowsStore from "@/store/useFlowsStore";
-import { useEffect } from "react";
-
 export default function Home() {
-  const updateFlows = useFlowsStore((state) => state.updateFlows);
-  const flows = useFlowsStore((state) => state.flows);
-  useEffect(() => {
-    const getflows = async () => {
-      const data = await readFlowsFromDatabase();
-      updateFlows(data);
-      console.log("called");
-    };
-
-    getflows().catch(console.error);
-  }, [updateFlows]);
-  return <SideNav flows={flows} />;
+  return <SideNav />;
 }
-
-// import SideNav from "@/components/SideNav";
-// import { readFlowsFromDatabase } from "@/lib/api-controllers";
-// import useFlowsStore from "@/store/useFlowsStore";
-// import { useEffect } from "react";
-
-// export default function Home() {
-//   return <SideNav />;
-// }
